Missing casting

+++ packages/hardhat/contracts/SculptureFactory.sol
@@ -137,7 +137,7 @@ contract Sculpture {
         if (_conservationData.conservation) {
             conservationData.conservationLabel = _conservationData.conservationLabel;
         } else {
-            conservationData.conservationLabel = SculptureLibrary.ConservationData.NONE;
+            conservationData.conservationLabel = uint8(SculptureLibrary.ConservationData.NONE);
         }
     }
